+++ packages/tools/src/tools/PlanarRotateTool.ts
@@ -1,14 +1,18 @@
-import { getEnabledElement, Types } from '@cornerstonejs/core';
-import { vec3 } from 'gl-matrix';
+import {
+  BaseVolumeViewport,
+  getEnabledElement,
+  Types,
+} from '@cornerstonejs/core';
+import { mat4, vec3 } from 'gl-matrix';
 import { BaseTool } from './base';
 import angleBetweenLines from '../utilities/math/angle/angleBetweenLines';
 import { PublicToolProps, ToolProps, EventTypes } from '../types';
 
 /**
- * The StackRotateTool is a tool that allows the user to rotate
- * the image on stack viewport by pressing the mouse click and dragging
+ * The PlanarRotateTool is a tool that allows the user to rotate
+ * the image by pressing the mouse click and dragging
  */
-class StackRotateTool extends BaseTool {
+class PlanarRotateTool extends BaseTool {
   static toolName;
   touchDragCallback: (evt: EventTypes.MouseDragEventType) => void;
   mouseDragCallback: (evt: EventTypes.MouseDragEventType) => void;
@@ -43,7 +47,7 @@ class StackRotateTool extends BaseTool {
       [centerWorld, currentPointWorld]
     );
 
-    const { viewPlaneNormal } = camera;
+    const { viewPlaneNormal, viewUp } = camera;
 
     const v1 = vec3.sub(vec3.create(), centerWorld, startPointWorld);
     const v2 = vec3.sub(vec3.create(), centerWorld, currentPointWorld);
@@ -54,12 +58,20 @@ class StackRotateTool extends BaseTool {
 
     if (Number.isNaN(angle)) return;
 
-    const { rotation } = (viewport as Types.IStackViewport).getProperties();
-    viewport.setProperties({ rotation: rotation + angle });
+    if (viewport instanceof BaseVolumeViewport) {
+      const rotAngle = (angle * Math.PI) / 180;
+      const rotMat = mat4.identity(new Float32Array(16));
+      mat4.rotate(rotMat, rotMat, rotAngle, viewPlaneNormal);
+      const rotatedViewUp = vec3.transformMat4(vec3.create(), viewUp, rotMat);
+      viewport.setCamera({ viewUp: rotatedViewUp as Types.Point3 });
+    } else {
+      const { rotation } = (viewport as Types.IStackViewport).getProperties();
+      viewport.setProperties({ rotation: rotation + angle });
+    }
 
     viewport.render();
   }
 }
 
-StackRotateTool.toolName = 'StackRotate';
-export default StackRotateTool;
+PlanarRotateTool.toolName = 'PlanarRotate';
+export default PlanarRotateTool;